Allow the use of the label and value if the source provides them.
Use the label for everything user-visible, but use the value in the
hidden select box and for comparisons. This supports the case when the
developer wants to show textual values to the user, but send a
programmatic value during form submission. Right now, the limitation is
that new values typed in by the user will be added to the array as a
string, whereas the values picked from the autocomplete list will be
added as a dictionary with label and value keys.