Fix the case where the user could add a new duplicate tag by fully typing it in.
For example, if the tagSource contains an entry with the label "base",
the user could select "base" from the dropdown and then type in "base"
quickly and hit Tab. I'm choosing to ignore this, since most likely they
meant to use the value from the drop-down and not to create a new tag
with no value. Unfortunately, it's impossible right now to detect the
reverse order of events (free-hand "base" first and then selecting
"base" from the list), since the new tag will not have a value.